code/requirements.txt
ultralytics
torch
opencv-python
matplotlib
Pillow
transformers
